% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMAT.R
\name{BERT_download}
\alias{BERT_download}
\title{DownLoad BERT models.}
\usage{
BERT_download(models)
}
\arguments{
\item{models}{Model names at \href{https://huggingface.co/models}{HuggingFace}.

For a full list of available BERT models, see
\url{https://huggingface.co/models?pipeline_tag=fill-mask&library=transformers}}
}
\value{
No return value.
Model files will be saved in "\%USERPROFILE\%/.cache/huggingface".
}
\description{
DownLoad BERT models to local cache folder "\%USERPROFILE\%/.cache/huggingface".
}
\examples{
\dontrun{
model.names = c("bert-base-uncased", "bert-base-cased")
BERT_download(model.names)
}

}
\seealso{
\code{\link{FMAT_load}}
}
