% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMAT.R
\name{BERT_info_date}
\alias{BERT_info_date}
\title{Scrape the earliest release date of BERT models.}
\usage{
BERT_info_date(models = NULL)
}
\arguments{
\item{models}{Model names at
\href{https://huggingface.co/models?pipeline_tag=fill-mask&library=transformers}{HuggingFace}.}
}
\value{
A data.table:
\itemize{
\item model name
\item earliest date (scraped from huggingface commit history)
}
}
\description{
Scrape the earliest release date of BERT models.
}
\examples{
\dontrun{
model.date = BERT_info_date()
# get all models from cache folder

one.model.date = FMAT:::get_model_date("bert-base-uncased")
# call the internal function to scrape a model
# that may not have been saved in cache folder
}

}
