% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMAT.R
\name{BERT_remove}
\alias{BERT_remove}
\title{Remove BERT models from local cache folder.}
\usage{
BERT_remove(models)
}
\arguments{
\item{models}{Model names.}
}
\value{
\code{NULL}.
}
\description{
Remove BERT models from local cache folder.
}
