% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMAT.R
\name{.}
\alias{.}
\title{A wrapper of \code{list()}.}
\usage{
.(...)
}
\arguments{
\item{...}{Objects (possibly named) passed to \code{\link{list}}.}
}
\value{
A list.
}
\description{
A simple version of the \code{\link{list}} function.
}
\examples{
.(Male=s("man, he, his"), Female=s("woman, she, her"))

}
