% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMAT.R
\docType{package}
\name{FMAT-package}
\alias{FMAT}
\alias{FMAT-package}
\title{FMAT: The Fill-Mask Association Test}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The Fill-Mask Association Test ('FMAT') \doi{10.1037/pspa0000396} is an integrative, probability-based social computing method using Masked Language Models to measure conceptual associations (e.g., attitudes, biases, stereotypes, social norms, cultural values) as propositional semantic representations in natural language. Supported language models include 'BERT' \doi{10.48550/arXiv.1810.04805} and its variants available at 'Hugging Face' \url{https://huggingface.co/models?pipeline_tag=fill-mask}. Methodological references and installation guidance are provided at \url{https://psychbruce.github.io/FMAT/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://psychbruce.github.io/FMAT/}
  \item Report bugs at \url{https://github.com/psychbruce/FMAT/issues}
}

}
\author{
\strong{Maintainer}: Han Wu Shuang Bao \email{baohws@foxmail.com} (\href{https://orcid.org/0000-0003-3043-710X}{ORCID})

}
\keyword{internal}
