% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMAT.R
\name{FMAT_query_bind}
\alias{FMAT_query_bind}
\title{Combine multiple query data.tables and renumber query ids.}
\usage{
FMAT_query_bind(...)
}
\arguments{
\item{...}{Query data.tables returned from \code{\link{FMAT_query}}.}
}
\value{
A data.table of queries and variables.
}
\description{
Combine multiple query data.tables and renumber query ids.
}
\examples{
FMAT_query_bind(
  FMAT_query(
    "[MASK] is {TARGET}.",
    MASK = .(Male="He", Female="She"),
    TARGET = .(Occupation=cc("a doctor, a nurse, an artist"))
  ),
  FMAT_query(
    "[MASK] occupation is {TARGET}.",
    MASK = .(Male="His", Female="Her"),
    TARGET = .(Occupation=cc("doctor, nurse, artist"))
  )
)

}
\seealso{
\code{\link{FMAT_load}}

\code{\link{FMAT_query}}

\code{\link{FMAT_run}}
}
