% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMAT.R
\name{ICC_models}
\alias{ICC_models}
\title{Intraclass correlation coefficient (ICC) of BERT models.}
\usage{
ICC_models(data, type = "agreement", unit = "average")
}
\arguments{
\item{data}{Raw data returned from \code{\link[=FMAT_run]{FMAT_run()}}.}

\item{type}{Interrater \code{"agreement"} (default) or \code{"consistency"}.}

\item{unit}{Reliability of \code{"average"} scores (default) or \code{"single"} scores.}
}
\value{
A data.table of ICC.
}
\description{
Interrater agreement of log probabilities (treated as "ratings"/rows) among BERT language models (treated as "raters"/columns), with both row and column as ("two-way") random effects.
}
