% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMAT.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{cc}
\alias{text_init}
\alias{text_model_download}
\alias{text_model_remove}
\alias{text_unmask}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{PsychWordVec}{\code{\link[PsychWordVec]{text_init}}, \code{\link[PsychWordVec]{text_model_download}}, \code{\link[PsychWordVec]{text_model_remove}}, \code{\link[PsychWordVec]{text_unmask}}}

  \item{bruceR}{\code{\link[bruceR]{cc}}}
}}

