% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMAT.R
\name{s}
\alias{s}
\title{Split a string (with separators) into a character vector.}
\usage{
s(x)
}
\arguments{
\item{x}{Character string.
Separators can be: \code{,} \code{;} \code{|} \code{\\n} \code{\\t}.}
}
\value{
Character vector.
}
\description{
Split a string (with separators) into a character vector.
}
\examples{
s("a, b, c, d, e")

}
