% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMAT.R
\name{summary.fmat}
\alias{summary.fmat}
\title{[S3 method] Summarize results of the FMAT.}
\usage{
\method{summary}{fmat}(fmat, ...)
}
\arguments{
\item{fmat}{A data.table (of new class \code{fmat})
returned by \code{\link{FMAT_run}}.}

\item{...}{Other arguments (currently not used).}
}
\value{
A data.table of summarized results.
}
\description{
Summarize the results of \emph{Log Probability Ratio} (LPR) for the FMAT.
}
\seealso{
\code{\link{FMAT_run}}
}
