% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMAT.R
\name{BERT_download}
\alias{BERT_download}
\title{Download and save BERT models to local cache folder.}
\usage{
BERT_download(models = NULL)
}
\arguments{
\item{models}{Model names at
\href{https://huggingface.co/models?pipeline_tag=fill-mask&library=transformers}{HuggingFace}.}
}
\value{
No return value.
}
\description{
Download and save BERT models to local cache folder "\%USERPROFILE\%/.cache/huggingface".
}
\examples{
\dontrun{
models = c("bert-base-uncased", "bert-base-cased")
BERT_download(models)

BERT_download()  # check downloaded models
}

}
\seealso{
\code{\link{BERT_vocab}}

\code{\link{FMAT_load}}
}
