% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMAT.R
\name{.}
\alias{.}
\title{A simple function equivalent to \code{list}.}
\usage{
.(...)
}
\arguments{
\item{...}{Named objects (usually character vectors for this package).}
}
\description{
A simple function equivalent to \code{list}.
}
\examples{
.(Male=cc("he, his"), Female=cc("she, her"))
list(Male=cc("he, his"), Female=cc("she, her"))  # the same

}
