% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMAT.R
\name{.}
\alias{.}
\title{A simple function equivalent to \code{list}.}
\usage{
.(...)
}
\arguments{
\item{...}{Named objects (usually character vectors for this package).}
}
\value{
A list of named objects.
}
\description{
A simple function equivalent to \code{list}.
}
\examples{
.(Male=c("he", "his"), Female=c("she", "her"))

}
\keyword{internal}
