% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMAT.R
\name{FMAT}
\alias{FMAT}
\title{The Fill-Mask Association Test (\code{FMAT_query()} & \code{FMAT_run()}).}
\usage{
FMAT(
  model = "bert-base-uncased",
  query = "[MASK] is {TARGET}.",
  MASK = .(),
  TARGET = .(),
  ATTRIB = .(),
  unmask.id = 1,
  parallel = FALSE
)
}
\arguments{
\item{query}{Query text (should be a character string/vector
with at least one \code{[MASK]} token).
Multiple queries would share the same
\code{MASK}, \code{TARGET}, and \code{ATTRIB}.
For multiple queries with different
\code{MASK}, \code{TARGET}, and/or \code{ATTRIB},
please use \code{\link{FMAT_query_bind}} to combine them.}

\item{MASK}{A named list of \code{[MASK]} target words.
Must be single words in the vocabulary of a certain masked language model.
Note that infrequent words usually do not exist in a model's vocabulary,
and in such a situation you may insert the words into the context by
specifying either \code{TARGET} or \code{ATTRIB}.}

\item{TARGET, ATTRIB}{A named list of Target/Attribute words or phrases.
If specified, then \code{query} must contain
\code{{TARGET}} and/or \code{{ATTRIB}} that would be
replaced by the words/phrases.}

\item{unmask.id}{If there are multiple \code{[MASK]} in \code{query},
this argument will be used to determine which one is to be unmasked.
Defaults to the 1st \code{[MASK]}.}
}
\description{
The Fill-Mask Association Test (\code{FMAT_query()} & \code{FMAT_run()}).
}
\examples{
models = c("bert-base-uncased",
           "bert-base-cased")

FMAT(
  models,
  "The {TARGET} has a [MASK] association with {ATTRIB}.",
  MASK = .(H="high", L="low"),
  TARGET = .(Flower=cc("rose, iris, lily"),
             Insect=cc("ant, cockroach, spider")),
  ATTRIB = .(Pos=cc("health, happiness, love, peace"),
             Neg=cc("death, sickness, hatred, disaster"))
)

}
\seealso{
\code{\link{FMAT_query}}

\code{\link{FMAT_query_bind}}

\code{\link{FMAT_run}}
}
