% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMAT.R
\name{FMAT_run}
\alias{FMAT_run}
\title{Run the mask filling pipeline on multiple models.}
\usage{
FMAT_run(
  models,
  data,
  progress = c("text", "time", "none"),
  parallel = FALSE,
  ncores = min(length(models), parallel::detectCores())
)
}
\arguments{
\item{models}{Language model(s):
\itemize{
  \item{Model names (usually the BERT-based models) at
   \href{https://huggingface.co/models}{HuggingFace}.}
  \item{A list of mask filling pipelines loaded by \code{\link{FMAT_load}}.

   * You will need to \strong{rerun} \code{\link{FMAT_load}}
   if you \strong{restart} the R session.}
}}

\item{data}{A data.table returned from
\code{\link{FMAT_query}} or \code{\link{FMAT_query_bind}}.}

\item{progress}{Show a progress bar:
\code{"text"} (default), \code{"time"}, \code{"none"}.}

\item{parallel}{Parallel processing. Defaults to \code{FALSE}.
If \code{TRUE}, then \code{models} must be model names
rather than from \code{\link{FMAT_load}}.

* For small-scale \code{data},
parallel processing would instead be \emph{slower}
because it takes time to create a parallel cluster.}

\item{ncores}{Number of CPU cores to be used in parallel processing.
Defaults to the minimum of the number of models and your CPU cores.}
}
\value{
A data.table (of new class \code{fmat}) appending \code{data}
with these new variables:
\itemize{
  \item{\code{model}: model name.}
  \item{\code{output}: complete sentence output with unmasked token.}
  \item{\code{token}: actual token to be filled in the blank mask
  (a note "out-of-vocabulary" will be added
  if the original word is not found in the model vocabulary).}
  \item{\code{prop}: (raw) conditional probability of the unmasked token
  given the provided context, estimated by the masked language model.

  * It is NOT SUGGESTED to directly interpret the raw probabilities
  because the \emph{contrast} between a pair of probabilities
  is more interpretable. See \code{\link{summary.fmat}}.}
}
}
\description{
Run the mask filling pipeline on multiple models.
}
\examples{
# Running the example requires the models downloaded
# You will need to rerun `FMAT_load` if you restart the R session
\donttest{
models = FMAT_load(c("bert-base-uncased", "bert-base-cased"))

dq = FMAT_query(
  c("[MASK] is {TARGET}.", "[MASK] works as {TARGET}."),
  MASK = .(Male="He", Female="She"),
  TARGET = .(Occupation=s("a doctor, a nurse, an artist"))
)
data1 = FMAT_run(models, dq)
summary(data1)

data2 = FMAT_run(
  models,
  FMAT_query(
    "The {TARGET} has a [MASK] association with {ATTRIB}.",
    MASK = .(H="high", L="low"),
    TARGET = .(Flower=s("rose, iris, lily"),
               Insect=s("ant, cockroach, spider")),
    ATTRIB = .(Pos=s("health, happiness, love, peace"),
               Neg=s("death, sickness, hatred, disaster"))
  ))
summary(data2)
}
}
\seealso{
\code{\link{FMAT_load}}

\code{\link{FMAT_query}}

\code{\link{FMAT_query_bind}}

\code{\link{summary.fmat}}
}
